test: rename to `test_main.py` for pytest compat
