+++ pysr/test/__init__.py
@@ -1,7 +1,7 @@
-from .test import runtests
 from .test_cli import get_runtests as get_runtests_cli
 from .test_dev import runtests as runtests_dev
 from .test_jax import runtests as runtests_jax
+from .test_main import runtests
 from .test_startup import runtests as runtests_startup
 from .test_torch import runtests as runtests_torch
