+++ pysr/test/test_main.py
